+++ Math/Vector.cpp
@@ -12,7 +12,7 @@ Vector::Vector()
 {
 }
 
-Vector::Vector(double x, double y, double z)
+Vector::Vector(float x, float y, float z)
     : m_x(x)
     , m_y(y)
     , m_z(z)
@@ -54,12 +54,12 @@ Vector Vector::operator-() const
     return Vector(-m_x, -m_y, -m_z);
 }
 
-Vector Vector::operator*(double rhs) const
+Vector Vector::operator*(float rhs) const
 {
     return Vector(rhs * m_x, rhs * m_y, rhs * m_z);
 }
 
-double Vector::dot(const Vector& rhs) const
+float Vector::dot(const Vector& rhs) const
 {
     return m_x * rhs.m_x + m_y * rhs.m_y + m_z * rhs.m_z;
 }
@@ -74,7 +74,7 @@ Vector Vector::cross(const Vector& a, const Vector& b)
     return a.cross(b);
 }
 
-Vector& Vector::operator*=(double rhs)
+Vector& Vector::operator*=(float rhs)
 {
     m_x *= rhs;
     m_y *= rhs;
@@ -82,7 +82,7 @@ Vector& Vector::operator*=(double rhs)
     return *this;
 }
 
-Vector& Vector::operator/=(double rhs)
+Vector& Vector::operator/=(float rhs)
 {
     m_x /= rhs;
     m_y /= rhs;
@@ -90,37 +90,37 @@ Vector& Vector::operator/=(double rhs)
     return *this;
 }
 
-Vector Vector::operator/(double rhs) const
+Vector Vector::operator/(float rhs) const
 {
     return Vector(m_x / rhs, m_y / rhs, m_z / rhs);
 }
 
-Vector operator*(double lhs, const Vector& rhs)
+Vector operator*(float lhs, const Vector& rhs)
 {
     return rhs * lhs;
 }
 
-double Vector::length() const
+float Vector::length() const
 {
     return std::sqrt(lengthSquared());
 }
 
-double Vector::lengthSquared() const
+float Vector::lengthSquared() const
 {
     return m_x * m_x + m_y * m_y + m_z * m_z;
 }
 
-double Vector::getX() const
+float Vector::getX() const
 {
     return m_x;
 }
 
-double Vector::getY() const
+float Vector::getY() const
 {
     return m_y;
 }
 
-double Vector::getZ() const
+float Vector::getZ() const
 {
     return m_z;
 }
